Refactor GPIO and convert functions, add clock module
Refactored GPIO and convert library functions for improved type safety and clarity. Added new clock.c and clock.h modules for STM32G031K8T6 clock configuration. Updated main.c to use GPIO toggling and improved pin initialization. Enhanced enum documentation and usage in gpio.h.

+++ Firmware/lib/convert.h
@@ -12,9 +12,9 @@
 /**
  * @brief Converts binary
  */
-uint32_t BinToDec(uint32_t bin);
+uint32_t BinToDec(uint64_t bin);
 
 /**
  * @brief Converts decimal to binary
  */
-uint32_t DecToBin(uint32_t dec);
+uint64_t DecToBin(uint32_t dec);
